a long is an integer, not a floating point number
git-svn-id: https://svn.mcs.anl.gov/repos/parallel-netcdf/trunk@247 c368ceb8-fd28-0410-a457-e46ee1d7cd4a

+++ test/nf_test/fortlib.c
@@ -191,7 +191,7 @@ FORTRAN_API double max_float_() {
 /* Else leave name alone */
 #endif
 FORTRAN_API double max_double_() {
-	return LONG_MAX;
+	return DBL_MAX;
 }
 
 #ifdef F77_NAME_UPPER
